# docs/swagger.yaml
basePath: /api/v1
definitions:
  sds-admin_internal_dto.CreateDomainRequest:
    properties:
      description:
        type: string
      domainName:
        type: string
      recursive:
        type: boolean
    required:
    - domainName
    type: object
...
  sds-admin_internal_dto.DomainResponse:
    properties:
      createdAt:
        type: string
      description:
        type: string
      domainName:
        type: string
      id:
        type: integer
      recursive:
        type: boolean
      status:
        type: string
      updatedAt:
        type: string
    type: object
  sds-admin_internal_dto.ErrorResponse:
    properties:
      code:
        type: integer
      detail:
        type: string
      message:
        type: string
    type: object
...
  sds-admin_internal_dto.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  sds-admin_internal_dto.UpdateDomainConfigRequest:
    properties:
      description:
        type: string
      recursive:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: SDS Admin API for domain management
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SDS Admin API
  version: "1.0"
paths:
  /domains:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all domains
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/sds-admin_internal_dto.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/sds-admin_internal_dto.DomainResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
      summary: List all domains
      tags:
      - domains
    post:
      consumes:
      - application/json
      description: Create a new domain with the provided information
      parameters:
      - description: Domain creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sds-admin_internal_dto.CreateDomainRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/sds-admin_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/sds-admin_internal_dto.DomainResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
      summary: Create a new domain
      tags:
      - domains
  /domains/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a domain by its ID (soft delete)
      parameters:
      - description: Domain ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
      summary: Delete a domain
      tags:
      - domains
    get:
      consumes:
      - application/json
      description: Retrieve a single domain by its ID
      parameters:
      - description: Domain ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/sds-admin_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/sds-admin_internal_dto.DomainResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
      summary: Get a domain by ID
      tags:
      - domains
    put:
      consumes:
      - application/json
      description: Update an existing domain with new information
      parameters:
      - description: Domain ID
        in: path
        name: id
        required: true
        type: integer
      - description: Domain update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sds-admin_internal_dto.CreateDomainRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/sds-admin_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/sds-admin_internal_dto.DomainResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
      summary: Update a domain
      tags:
      - domains
  /domains/{id}/config:
    patch:
      consumes:
      - application/json
      description: Update domain config (description and recursive)
      parameters:
      - description: Domain ID
        in: path
        name: id
        required: true
        type: integer
      - description: Domain config update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sds-admin_internal_dto.UpdateDomainConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/sds-admin_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/sds-admin_internal_dto.DomainResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
      summary: Update domain config
      tags:
      - domains
  /domains/{id}/disable:
    post:
      consumes:
      - application/json
      description: Disable a domain by setting its status to inactive
      parameters:
      - description: Domain ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/sds-admin_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/sds-admin_internal_dto.DomainResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
      summary: Disable a domain
      tags:
      - domains
  /domains/{id}/enable:
    post:
      consumes:
      - application/json
      description: Enable a domain by setting its status to active
      parameters:
      - description: Domain ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/sds-admin_internal_dto.Response'
            - properties:
                data:
                  $ref: '#/definitions/sds-admin_internal_dto.DomainResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/sds-admin_internal_dto.ErrorResponse'
      summary: Enable a domain
      tags:
      - domains
...
securityDefinitions:
  ApiKeyAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"